requirements.txt
Django==4.2
django-debug-toolbar==4.1.0
django-environ==0.11.2
pillow==11.1.0
psycopg2-binary==2.9.10
djangorestframework==3.15.2
django-filter==24.3
djangorestframework_simplejwt==5.4.0
django-cors-headers==4.7.0
django-redis==5.4.0
drf-nested-routers==0.94.1
celery==5.5.1
django-celery-beat==2.7.0


